rename serialize spec file
