+++ src/presentation.ts
@@ -26,7 +26,7 @@ import {
 import { assert, isSubclass, zip } from './util.ts';
 import { generateContext, computeContext } from './context.ts';
 import { NestedProvable } from './nested.ts';
-import { serializeSpec, serializeInputContext } from './serialize.ts';
+import { serializeSpec, serializeInputContext } from './serialize-spec.ts';
 import { deserializeSpec, deserializeInputContext } from './deserialize.ts';
 import {
   deserializeNestedProvableValue,
