+++ src/serialize-spec.ts
